Adds unit tests for decodeNullDataIPNS
